+++ hibike/hibike_process.py
@@ -1,25 +1,27 @@
 """
 The main Hibike process.
 """
-from collections import namedtuple
+import asyncio
 import glob
-import multiprocessing
 import os
-import queue
-import random
-import threading
-import time
 import sys
+import random
+import time
 
-#from PieCentral.runtime.runtimeUtil import *
-# from runtimeUtil import *
-
-# pylint: disable=import-error
 import hibike_message as hm
-import serial
+try:
+    import hibike_packet
+    USING_PACKET_EXTENSION = True
+except ImportError:
+    USING_PACKET_EXTENSION = False
 
+import serial_asyncio
+import aioprocessing
+import aiofiles
+import uvloop
 __all__ = ["hibike_process"]
 
+asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
 # .04 milliseconds sleep is the same frequency we subscribe to devices at
 BATCH_SLEEP_TIME = .04
@@ -28,174 +30,243 @@ IDENTIFY_TIMEOUT = 1
 # Time in seconds to wait between checking for new devices
 # and cleaning up old ones.
 HOTPLUG_POLL_INTERVAL = 1
+# Whether to use profiling or not. On the BBB, profiling adds a significant overhead (~30%).
+USE_PROFILING = False
+# Where to output profiling statistics. By default, this is in Callgrind format
+PROFILING_OUTPUT_FILE = "func_stats"
+# The time period to take measurements over, in seconds
+PROFILING_PERIOD = 60
+PAUSE_QUEUE_SIZE = 10
+RESUME_QUEUE_SIZE = 2
 
-
-def get_working_serial_ports(excludes=()):
+def scan_for_serial_ports():
     """
-    Scan for open COM ports, except those in EXCLUDES.
-
-    Returns:
-        A list of serial port objects (`serial.Serial`) and port names.
+    Scan for serial ports that look like an Arduino.
     """
-    excludes = set(excludes)
     # Last command is included so that it's compatible with OS X Sierra
     # Note: If you are running OS X Sierra, do not access the directory through vagrant ssh
     # Instead access it through Volumes/vagrant/PieCentral
-    ports = set(glob.glob("/dev/ttyACM*") + glob.glob("/dev/ttyUSB*")
-                + glob.glob("/dev/tty.usbmodem*"))
-    ports.difference_update(excludes)
+    return set(glob.glob("/dev/ttyACM*") +\
+               glob.glob("/dev/ttyUSB*") +\
+               glob.glob("/dev/tty.usbmodem*"))
+
+
+async def get_working_serial_ports(event_loop, excludes=()):
+    """
+    Scan for open COM ports, except those in `excludes`.
+
+    Returns:
+        A list of port names.
+    """
+    excludes = set(excludes)
+    ports = await event_loop.run_in_executor(None, scan_for_serial_ports)
     try:
         virtual_device_config_file = os.path.join(os.path.dirname(__file__), "virtual_devices.txt")
-        ports.update(open(virtual_device_config_file, "r").read().split())
+        async with aiofiles.open(virtual_device_config_file, loop=event_loop) as f:
+            contents = await f.read()
+            ports.update(contents.split())
     except IOError:
         pass
-
-    serials = []
-    port_names = []
-    for port in ports:
-        try:
-            serials.append(serial.Serial(port, 115200))
-            port_names.append(port)
-        except serial.serialutil.SerialException:
-            print("Cannot Open Serial Port: " + str(port))
-    return serials, port_names
+    ports.difference_update(excludes)
+    return list(ports)
 
 
-def identify_smart_sensors(serial_conns):
+async def hotplug_async(devices, batched_data, error_queue, state_queue, event_loop):
     """
-    Given a list of serial port connections, figure out which
-    contain smart sensors.
-
-    Returns:
-        A map of serial port names to UIDs.
+    Scan for new devices on serial ports and automatically spin them up.
     """
-    def recv_subscription_response(conn, uid_queue, stop_event):
+    pending = set()
+    def protocol_factory():
         """
-        Place received subscription response UIDs from CONN into UID_QUEUE,
-        stopping when STOP_EVENT is set.
+        Create a `SmartSensorProtocol` with necessary parameters filled in.
         """
-        try:
-            for packet in hm.blocking_read_generator(conn, stop_event):
-                msg_type = packet.get_message_id()
-                if msg_type == hm.MESSAGE_TYPES["SubscriptionResponse"]:
-                    _, _, uid = hm.parse_subscription_response(packet)
-                    uid_queue.put(uid)
-        except serial.SerialException:
-            pass
+        return SmartSensorProtocol(devices, batched_data, error_queue,
+                                   state_queue, event_loop, pending)
 
-
-    device_map = {}
-    candidates = []
-    for conn in serial_conns:
-        old_timeout = conn.write_timeout
-        conn.write_timeout = IDENTIFY_TIMEOUT
-        try:
-            hm.send(conn, hm.make_ping())
-        except serial.SerialTimeoutException:
-            continue
-        finally:
-            conn.write_timeout = old_timeout
-        maybe_device = namedtuple("MaybeDevice", ["serial_conn", "queue", "event", "thread"])
-        maybe_device.queue = queue.Queue()
-        maybe_device.event = threading.Event()
-        maybe_device.serial_conn = conn
-        maybe_device.thread = threading.Thread(target=recv_subscription_response,
-                                               args=(conn, maybe_device.queue, maybe_device.event))
-        candidates.append(maybe_device)
-    for cand in candidates:
-        cand.thread.start()
-    for cand in candidates:
-        try:
-            uid = cand.queue.get(block=True, timeout=IDENTIFY_TIMEOUT)
-            device_map[cand.serial_conn.name] = uid
-            # Shut device up
-            hm.send(cand.serial_conn, hm.make_subscription_request(uid, [], 0))
-        except queue.Empty:
-            pass
-    for cand in candidates:
-        cand.event.set()
-        cand.thread.join()
-    return device_map
-
-
-def spin_up_device(serial_port, uid, state_queue, batched_data, error_queue):
-    """
-    Spin up a device with a given UID on SERIAL_PORT.
-
-    Returns:
-        The new device.
-    """
-    pack = namedtuple("Threadpack", ["read_thread", "write_thread",
-                                     "write_queue", "serial_port", "instance_id"])
-    pack.write_queue = queue.Queue()
-    pack.serial_port = serial_port
-    pack.write_thread = threading.Thread(target=device_write_thread,
-                                         args=(serial_port, pack.write_queue))
-    pack.read_thread = threading.Thread(target=device_read_thread,
-                                        args=(uid, pack, error_queue,
-                                              state_queue, batched_data))
-    # This is an ID that does not persist across disconnects,
-    # so that we can tell when a device has been reconnected.
-    pack.instance_id = random.getrandbits(128)
-    pack.write_thread.start()
-    pack.read_thread.start()
-    return pack
-
-
-def hotplug(devices, state_queue, batched_data, error_queue):
-    """
-    Remove disconnected devices and scan for new ones.
-    """
-    clean_up_queue = queue.Queue()
-    clean_up_thread = threading.Thread(target=clean_up_devices, args=(clean_up_queue, ))
-    clean_up_thread.start()
     while True:
-        time.sleep(HOTPLUG_POLL_INTERVAL)
-        scan_for_new_devices(devices, state_queue, batched_data, error_queue)
-        remove_disconnected_devices(error_queue, devices, clean_up_queue, state_queue)
+        await asyncio.sleep(HOTPLUG_POLL_INTERVAL, loop=event_loop)
+        port_names = set([dev.transport.serial.name for dev in devices.values()\
+                          if dev.transport is not None and dev.transport.serial is not None])
+        port_names.update(pending)
+        new_serials = await get_working_serial_ports(event_loop, port_names)
+        for port in new_serials:
+            try:
+                pending.add(port)
+                await serial_asyncio.create_serial_connection(event_loop, protocol_factory, port,
+                                                              baudrate=115200)
+            except serial_asyncio.serial.SerialException:
+                pass
+        await remove_disconnected_devices(error_queue, devices, state_queue, event_loop)
 
 
-def scan_for_new_devices(existing_devices, state_queue, batched_data, error_queue):
+class SmartSensorProtocol(asyncio.Protocol):
     """
-    Find devices that are on serial ports not in EXISTING_DEVICES, and add
-    any that have been found to it.
+    Handle communication over serial with a smart sensor.
+
+    :param dict devices: Mapping from UIDs to devices
+    :param dict batched_data: Mapping from UIDs to device state
+    :param asyncio.Queue error_queue: Info about disconnects goes here
+    :param aioprocessing.Queue state_queue: Connection to StateManager
+    :param event_loop: The event loop
+    :param set pending: Set of serial connections that may or may not
+    have devices on them.
     """
-    ports, names = get_working_serial_ports(map(lambda d: d.serial_port.name,
-                                                existing_devices.values()))
-    sensors = identify_smart_sensors(ports)
-    for (ser, uid) in sensors.items():
-        idx = names.index(ser)
-        port = ports[idx]
-        pack = spin_up_device(port, uid, state_queue, batched_data, error_queue)
-        existing_devices[uid] = pack
-        # Tell the device to start sending data
-        pack.write_queue.put(("ping", []))
-        pack.write_queue.put(("subscribe", [1, 0, []]))
+    PACKET_BOUNDARY = bytes([0])
+    __slots__ = ("uid", "write_queue", "batched_data", "read_queue", "error_queue",
+                 "state_queue", "instance_id", "transport", "_ready", "serial_buf")
+    # pylint: disable=too-many-arguments
+    def __init__(self, devices, batched_data, error_queue, state_queue, event_loop, pending: set):
+        # We haven't found out what our UID is yet
+        self.uid = None
+
+        self.write_queue = asyncio.Queue(loop=event_loop)
+        self.batched_data = batched_data
+        self.read_queue = asyncio.Queue(loop=event_loop)
+        self.error_queue = error_queue
+        self.state_queue = state_queue
+        self.instance_id = random.getrandbits(128)
+
+        self.transport = None
+        self._ready = asyncio.Event(loop=event_loop)
+        if USING_PACKET_EXTENSION:
+            # pylint: disable=no-member
+            self.serial_buf = hibike_packet.RingBuffer()
+        else:
+            self.serial_buf = bytearray()
+
+        event_loop.create_task(self.register_sensor(event_loop, devices, pending))
+        event_loop.create_task(self.send_messages())
+        event_loop.create_task(self.recv_messages())
+
+    async def register_sensor(self, event_loop, devices, pending):
+        """
+        Try to get our UID from the sensor and register it with `hibike_process`.
+        """
+        await self._ready.wait()
+        hm.send(self.transport, hm.make_ping())
+        await asyncio.sleep(IDENTIFY_TIMEOUT, loop=event_loop)
+        if self.uid is None:
+            self.quit()
+        else:
+            hm.send(self.transport, hm.make_ping())
+            hm.send(self.transport,
+                    hm.make_subscription_request(hm.uid_to_device_id(self.uid), [], 0))
+            devices[self.uid] = self
+        pending.remove(self.transport.serial.name)
+
+    async def send_messages(self):
+        """
+        Send messages in the queue to the sensor.
+        """
+        await self._ready.wait()
+        while not self.transport.is_closing():
+            instruction, args = await self.write_queue.get()
+            if instruction == "ping":
+                hm.send(self.transport, hm.make_ping())
+            elif instruction == "subscribe":
+                uid, delay, params = args
+                hm.send(self.transport,
+                        hm.make_subscription_request(hm.uid_to_device_id(uid),
+                                                     params, delay))
+            elif instruction == "read":
+                uid, params = args
+                hm.send(self.transport, hm.make_device_read(hm.uid_to_device_id(uid), params))
+            elif instruction == "write":
+                uid, params_and_values = args
+                hm.send(self.transport, hm.make_device_write(hm.uid_to_device_id(uid),
+                                                             params_and_values))
+            elif instruction == "disable":
+                hm.send(self.transport, hm.make_disable())
+            elif instruction == "heartResp":
+                uid = args[0]
+                hm.send(self.transport, hm.make_heartbeat_response(self.read_queue.qsize()))
+
+    async def recv_messages(self):
+        """
+        Process received messages.
+        """
+        await self._ready.wait()
+        while not self.transport.is_closing():
+            if self.read_queue.qsize() >= PAUSE_QUEUE_SIZE:
+                self.transport.pause_reading()
+            if self.read_queue.qsize() <= RESUME_QUEUE_SIZE:
+                self.transport.resume_reading()
+            packet = await self.read_queue.get()
+            message_type = packet.get_message_id()
+            if message_type == hm.MESSAGE_TYPES["SubscriptionResponse"]:
+                params, delay, uid = hm.parse_subscription_response(packet)
+                self.uid = uid
+                await self.state_queue.coro_put(("device_subscribed", [uid, delay, params]))
+            elif message_type == hm.MESSAGE_TYPES["DeviceData"]:
+                # This is kind of a hack, but it allows us to use `recv_messages` for
+                # detecting new smart sensors as well as reading from known ones.
+                if self.uid is not None:
+                    params_and_values = hm.parse_device_data(packet, hm.uid_to_device_id(self.uid))
+                    self.batched_data[uid] = params_and_values
+            elif message_type == hm.MESSAGE_TYPES["HeartBeatRequest"]:
+                if self.uid is not None:
+                    self.write_queue.put_nowait(("heartResp", [self.uid]))
+
+    def connection_made(self, transport):
+        self.transport = transport
+        self._ready.set()
+
+    def quit(self):
+        """
+        Stop processing packets and close the serial connection.
+        """
+        self.transport.abort()
+
+    if USING_PACKET_EXTENSION:
+        def data_received(self, data):
+            self.serial_buf.extend(data)
+            # pylint: disable=no-member
+            maybe_packet = hibike_packet.process_buffer(self.serial_buf)
+            if maybe_packet is not None:
+                message_id, payload = maybe_packet
+                message = hm.HibikeMessage(message_id, payload)
+                self.read_queue.put_nowait(message)
+    else:
+        def data_received(self, data):
+            self.serial_buf.extend(data)
+            zero_loc = self.serial_buf.find(self.PACKET_BOUNDARY)
+            if zero_loc != -1:
+                self.serial_buf = self.serial_buf[zero_loc:]
+                packet = hm.parse_bytes(self.serial_buf)
+                if packet != None:
+                    # Chop off a byte so we don't output this packet again
+                    self.serial_buf = self.serial_buf[1:]
+                    self.read_queue.put_nowait(packet)
+                elif self.serial_buf.count(self.PACKET_BOUNDARY) > 1:
+                    # If there's another packet in the buffer
+                    # we can safely jump to it for the next iteration
+                    new_packet = self.serial_buf[1:].find(self.PACKET_BOUNDARY) + 1
+                    self.serial_buf = self.serial_buf[new_packet:]
+
+    def connection_lost(self, exc):
+        if self.uid is not None:
+            error = Disconnect(uid=self.uid, instance_id=self.instance_id, accessed=False)
+            self.error_queue.put_nowait(error)
 
 
-def clean_up_devices(device_queue):
+class Disconnect:
     """
-    Clean up associated resources of devices in the queue.
-
-    Closing a serial port can take a very long time (30 seconds or more).
-    It's best to spin this function off into its own thread,
-    so that you're not blocked on reclaiming resources.
+    Information about a device disconnect.
     """
-    while True:
-        device = device_queue.get()
-        device.serial_port.close()
-        device.read_thread.join()
-        device.write_thread.join()
+    def __init__(self, uid, instance_id, accessed):
+        self.uid = uid
+        self.instance_id = instance_id
+        self.accessed = accessed
 
 
-def remove_disconnected_devices(error_queue, devices, clean_up_queue, state_queue):
+async def remove_disconnected_devices(error_queue, devices, state_queue, event_loop):
     """
-    Clean up any disconnected devices in ERROR_QUEUE.
+    Clean up any disconnected devices in `error_queue`.
     """
     next_time_errors = []
     while True:
         try:
-            error = error_queue.get(block=False)
+            error = error_queue.get_nowait()
             pack = devices[error.uid]
             if not error.accessed:
                 # Wait until the next cycle to make sure it's disconnected
@@ -206,250 +277,126 @@ def remove_disconnected_devices(error_queue, devices, clean_up_queue, state_queu
                 # The device has reconnected in the meantime
                 continue
             uid = error.uid
-            pack = devices[uid]
             del devices[uid]
-            clean_up_queue.put(pack)
-            state_queue.put(("device_disconnected", [uid]))
-        except queue.Empty:
+            await state_queue.coro_put(("device_disconnected", [uid]), loop=event_loop)
+        except asyncio.QueueEmpty:
             for err in next_time_errors:
-                error_queue.put(err)
+                error_queue.put_nowait(err)
             return
 
-# pylint: disable=too-many-branches, too-many-locals
-# pylint: disable=too-many-arguments, unused-argument
+
+async def batch_data(sensor_values, state_queue, event_loop):
+    """
+    Periodically send sensor values to `StateManager`.
+    """
+    while True:
+        await asyncio.sleep(BATCH_SLEEP_TIME, loop=event_loop)
+        await state_queue.coro_put(("device_values", [sensor_values]), loop=event_loop)
+
+
+async def print_profiler_stats(event_loop, time_delay):
+    """
+    Print profiler statistics after a number of seconds.
+    """
+    try:
+        import yappi
+    except ImportError:
+        return
+    await asyncio.sleep(time_delay, loop=event_loop)
+    print("Printing profiler stats")
+    yappi.get_func_stats().print_all(out=sys.stdout,
+                                     columns={0: ("name", 60), 1: ("ncall", 5), 2: ("tsub", 8),
+                                              3: ("ttot", 8), 4: ("tavg", 8)})
+    yappi.get_func_stats().save(PROFILING_OUTPUT_FILE, type="callgrind")
+    yappi.get_thread_stats().print_all()
+
+
+class QueueContext:
+    """
+    Stub to force aioprocessing to use an existing queue.
+    """
+    def __init__(self, queue):
+        self._queue = queue
+
+    # pylint: disable=invalid-name
+    def Queue(self, _size):
+        return self._queue
+
+
 def hibike_process(bad_things_queue, state_queue, pipe_from_child):
     """
     Run the main hibike processs.
     """
-    serials, serial_names = get_working_serial_ports()
-    smart_sensors = identify_smart_sensors(serials)
+    pipe_from_child = aioprocessing.AioConnection(pipe_from_child)
+    # By default, AioQueue instantiates a new Queue object, but we
+    # don't want that.
+    state_queue = aioprocessing.AioQueue(context=QueueContext(state_queue))
+    bad_things_queue = aioprocessing.AioQueue(context=QueueContext(bad_things_queue))
+
     devices = {}
-
     batched_data = {}
-    error_queue = queue.Queue()
+    event_loop = asyncio.get_event_loop()
+    error_queue = asyncio.Queue(loop=event_loop)
 
-    for (ser, uid) in smart_sensors.items():
-        index = serial_names.index(ser)
-        serial_port = serials[index]
-        pack = spin_up_device(serial_port, uid, state_queue, batched_data, error_queue)
-        devices[uid] = pack
+    event_loop.create_task(batch_data(batched_data, state_queue, event_loop))
+    event_loop.create_task(hotplug_async(devices, batched_data, error_queue,
+                                         state_queue, event_loop))
+    event_loop.create_task(dispatch_instructions(devices, bad_things_queue, state_queue,
+                                                 pipe_from_child, event_loop))
+    # start event loop
+    if USE_PROFILING:
+        try:
+            import yappi
+            yappi.start()
+            event_loop.create_task(print_profiler_stats(event_loop, PROFILING_PERIOD))
+        except ImportError:
+            print("Unable to import profiler. Make sure you installed with the '--dev' flag.")
 
-    batch_thread = threading.Thread(target=batch_data, args=(batched_data, state_queue))
-    batch_thread.start()
-    hotplug_thread = threading.Thread(target=hotplug,
-                                      args=(devices, state_queue, batched_data, error_queue))
-    hotplug_thread.start()
+    event_loop.run_forever()
 
-    # Pings all devices and tells them to stop sending data
-    for pack in devices.values():
-        pack.write_queue.put(("ping", []))
-        pack.write_queue.put(("subscribe", [1, 0, []]))
 
-    # the main thread reads instructions from statemanager and
-    # forwards them to the appropriate device write threads
-
+async def dispatch_instructions(devices, bad_things_queue, state_queue,
+                                pipe_from_child, event_loop):
+    """
+    Respond to instructions from `StateManager`.
+    """
     path = os.path.dirname(os.path.abspath(__file__))
     parent_path = path.rstrip("hibike")
     runtime = os.path.join(parent_path, "runtime")
     sys.path.insert(1, runtime)
+    # Pylint doesn't understand our import shenanigans
+    # pylint: disable=import-error
     import runtimeUtil
 
     while True:
-        instruction, args = pipe_from_child.recv()
+        instruction, args = await pipe_from_child.coro_recv(loop=event_loop)
         try:
             if instruction == "enumerate_all":
                 for pack in devices.values():
-                    pack.write_queue.put(("ping", []))
+                    pack.write_queue.put_nowait(("ping", []))
             elif instruction == "subscribe_device":
                 uid = args[0]
-                if uid in devices:
-                    devices[uid].write_queue.put(("subscribe", args))
+                devices[uid].write_queue.put_nowait(("subscribe", args))
             elif instruction == "write_params":
                 uid = args[0]
-                if uid in devices:
-                    devices[uid].write_queue.put(("write", args))
+                devices[uid].write_queue.put_nowait(("write", args))
             elif instruction == "read_params":
                 uid = args[0]
-                if uid in devices:
-                    devices[uid].write_queue.put(("read", args))
+                devices[uid].write_queue.put_nowait(("read", args))
             elif instruction == "disable_all":
                 for pack in devices.values():
-                    pack.write_queue.put(("disable", []))
+                    pack.write_queue.put_nowait(("disable", []))
             elif instruction == "timestamp_down":
-                timestamp = time.perf_counter()
+                timestamp = time.time()
                 args.append(timestamp)
-                state_queue.put(("timestamp_up", args))
+                await state_queue.coro_put(("timestamp_up", args), loop=event_loop)
         except KeyError as e:
-            bad_things_queue.put(runtimeUtil.BadThing(
+            await bad_things_queue.coro_put(runtimeUtil.BadThing(
                 sys.exc_info(),
                 str(e),
                 event=runtimeUtil.BAD_EVENTS.HIBIKE_NONEXISTENT_DEVICE))
         except TypeError as e:
-            bad_things_queue.put(runtimeUtil.BadThing(
+            await bad_things_queue.coro_put(runtimeUtil.BadThing(
                 sys.exc_info(),
                 str(e),
                 event=runtimeUtil.BAD_EVENTS.HIBIKE_INSTRUCTION_ERROR))
-
-
-def device_write_thread(ser, instr_queue):
-    """
-    Send packets to SER based on instructions from INSTR_QUEUE.
-    """
-    try:
-        while True:
-            instruction, args = instr_queue.get()
-
-            if instruction == "ping":
-                hm.send(ser, hm.make_ping())
-            elif instruction == "subscribe":
-                uid, delay, params = args
-                hm.send(ser, hm.make_subscription_request(hm.uid_to_device_id(uid), params, delay))
-            elif instruction == "read":
-                uid, params = args
-                hm.send(ser, hm.make_device_read(hm.uid_to_device_id(uid), params))
-            elif instruction == "write":
-                uid, params_and_values = args
-                hm.send(ser, hm.make_device_write(hm.uid_to_device_id(uid), params_and_values))
-            elif instruction == "disable":
-                hm.send(ser, hm.make_disable())
-            elif instruction == "heartResp":
-                uid = args[0]
-                hm.send(ser, hm.make_heartbeat_response())
-    except serial.SerialException:
-        # Device has disconnected
-        pass
-
-
-def device_read_thread(uid, pack, error_queue, state_queue, batched_data):
-    """
-    Read packets from SER and update queues and BATCHED_DATA accordingly.
-    """
-    ser = pack.serial_port
-    instruction_queue = pack.write_queue
-    try:
-        while True:
-            for packet in hm.blocking_read_generator(ser):
-                message_type = packet.get_message_id()
-                if message_type == hm.MESSAGE_TYPES["SubscriptionResponse"]:
-                    params, delay, uid = hm.parse_subscription_response(packet)
-                    state_queue.put(("device_subscribed", [uid, delay, params]))
-                elif message_type == hm.MESSAGE_TYPES["DeviceData"]:
-                    params_and_values = hm.parse_device_data(packet, hm.uid_to_device_id(uid))
-                    batched_data[uid] = params_and_values
-                elif message_type == hm.MESSAGE_TYPES["HeartBeatRequest"]:
-                    instruction_queue.put(("heartResp", [uid]))
-    except serial.SerialException:
-        error = namedtuple("Disconnect", ["uid", "instance_id", "accessed"])
-        error.uid = uid
-        error.instance_id = pack.instance_id
-        error.accessed = False
-        error_queue.put(error)
-
-def batch_data(data, state_queue):
-    """
-    Write out DATA to STATE_QUEUE periodically.
-    """
-    while True:
-        time.sleep(BATCH_SLEEP_TIME)
-        state_queue.put(("device_values", [data]))
-
-#############
-## TESTING ##
-#############
-# pylint: disable=invalid-name
-if __name__ == "__main__":
-    # helper functions so we can spawn threads that try to read/write to hibike_devices periodically
-    def set_interval_sequence(functions, sec):
-        """
-        Create a thread that executes FUNCTIONS after SEC seconds.
-        """
-        def func_wrapper():
-            """
-            Execute the next function in FUNCTIONS after SEC seconds.
-
-            Cycles through all functions.
-            """
-            set_interval_sequence(functions[1:] + functions[:1], sec)
-            functions[0]()
-        t = threading.Timer(sec, func_wrapper)
-        t.start()
-        return t
-
-    def make_send_write(pipe_to_child, uid, params_and_values):
-        """
-        Create a function that sends UID and PARAMS_AND_VALUES
-        to PIPE_TO_CHILD.
-        """
-        def helper():
-            """
-            Helper function.
-            """
-            pipe_to_child.send(["write_params", [uid, params_and_values]])
-        return helper
-
-    to_child, from_child = multiprocessing.Pipe()
-    main_error_queue = multiprocessing.Queue()
-    main_state_queue = multiprocessing.Queue()
-    newProcess = multiprocessing.Process(target=hibike_process,
-                                         name="hibike_sim",
-                                         args=[main_error_queue, main_state_queue, from_child])
-    newProcess.daemon = True
-    newProcess.start()
-    to_child.send(["enumerate_all", []])
-    uids = set()
-    while True:
-        print("waiting for command")
-        command, main_args = main_state_queue.get()
-        if command == "device_subscribed":
-            dev_uid = main_args[0]
-            if dev_uid not in uids:
-                uids.add(dev_uid)
-                if hm.DEVICES[hm.uid_to_device_id(dev_uid)]["name"] == "YogiBear":
-                    set_interval_sequence([
-                        make_send_write(to_child, dev_uid, [("duty_cycle", 0)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", 0.5)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", 1.0)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", 0)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", -0.5)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", -1.0)]),
-                        make_send_write(to_child, dev_uid, [("duty_cycle", 0)])
-                        ], 0.75)
-                elif hm.DEVICES[hm.uid_to_device_id(dev_uid)]["name"] == "ServoControl":
-                    set_interval_sequence([
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 1), ("enable0", False),
-                                         ("servo1", 21), ("enable1", True),
-                                         ("servo2", 30), ("enable2", True),
-                                         ("servo3", 8), ("enable3", True)]),
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 5), ("enable0", False),
-                                         ("servo1", 5), ("enable1", True),
-                                         ("servo2", 5), ("enable2", True),
-                                         ("servo3", 5), ("enable3", False)]),
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 1), ("enable0", True),
-                                         ("servo1", 26), ("enable1", True),
-                                         ("servo2", 30), ("enable2", False),
-                                         ("servo3", 17), ("enable3", True)]),
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 13), ("enable0", False),
-                                         ("servo1", 7), ("enable1", False),
-                                         ("servo2", 24), ("enable2", True),
-                                         ("servo3", 10), ("enable3", True)]),
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 27), ("enable0", True),
-                                         ("servo1", 2), ("enable1", False),
-                                         ("servo2", 3), ("enable2", False),
-                                         ("servo3", 14), ("enable3", False)]),
-                        make_send_write(to_child, dev_uid,
-                                        [("servo0", 20), ("enable0", True),
-                                         ("servo1", 12), ("enable1", False),
-                                         ("servo2", 20), ("enable2", False),
-                                         ("servo3", 29), ("enable3", True)]),
-                        ], 1)
-                parameters = []
-                for param in hm.DEVICES[hm.uid_to_device_id(dev_uid)]["params"]:
-                    parameters.append(param["name"])
-                to_child.send(["subscribe_device", [dev_uid, 10, parameters]])
-        elif command == "device_values":
-            print("%10.2f, %s" % (time.time(), str(main_args)))
