+++ hibike/virtual_device.py
@@ -8,17 +8,182 @@ $ socat -d -d pty,raw,echo=0 pty,raw,echo=0
 2016/09/20 21:29:03 socat[4165] N starting data transfer loop with FDs [3,3] and [5,5]
 $ python3.5 virtual_device.py -d LimitSwitch -p /dev/pts/26
 """
-import random
-import time
 import argparse
+import asyncio
+import json
+import random
 import struct
+import time
 
 # pylint: disable=import-error
-import serial
+import serial_asyncio
 import hibike_message as hm
 
-# pylint: disable=too-many-statements, too-many-locals, too-many-branches
-# pylint: disable=unused-variable
+
+# Format of default values storage:
+# {"DeviceName": {"param1": value1, "param2": value2}}
+DEFAULT_VALUES_FILE = "virtual_device_defaults.json"
+with open(DEFAULT_VALUES_FILE) as f:
+    DEFAULT_VALUES = json.load(f)
+
+
+class VirtualDevice(asyncio.Protocol):
+    """
+    A fake Hibike smart sensor.
+    """
+    HEARTBEAT_DELAY_MS = 100
+    PACKET_BOUNDARY = bytes([0])
+    def __init__(self, uid, event_loop, verbose=False):
+        self.uid = uid
+        self.event_loop = event_loop
+        self._ready = asyncio.Event(loop=event_loop)
+        self.serial_buf = bytearray()
+        self.read_queue = asyncio.Queue(loop=event_loop)
+        self.verbose = verbose
+
+        self.update_time = 0
+        self.delay = 0
+        self.transport = None
+
+        self.response_map = {
+            hm.MESSAGE_TYPES["Ping"]: self._process_ping,
+            hm.MESSAGE_TYPES["SubscriptionRequest"]: self._process_sub_request,
+            hm.MESSAGE_TYPES["DeviceRead"]: self._process_device_read,
+            hm.MESSAGE_TYPES["DeviceWrite"]: self._process_device_write,
+            hm.MESSAGE_TYPES["Disable"]: self._process_disable,
+            hm.MESSAGE_TYPES["HeartBeatResponse"]: self._process_heartbeat_response,
+        }
+        self.param_values = DEFAULT_VALUES[hm.uid_to_device_name(uid)]
+        self.subscribed_params = set()
+
+        event_loop.create_task(self.process_messages())
+        event_loop.create_task(self.send_subscribed_params())
+        event_loop.create_task(self.request_heartbeats())
+
+    def data_received(self, data):
+        self.serial_buf.extend(data)
+        zero_loc = self.serial_buf.find(bytes([0]))
+        if zero_loc != -1:
+            self.serial_buf = self.serial_buf[zero_loc:]
+            packet = hm.parse_bytes(self.serial_buf)
+            if packet != None:
+                self.serial_buf = self.serial_buf[1:]
+                self.read_queue.put_nowait(packet)
+            elif self.serial_buf.count(self.PACKET_BOUNDARY) > 1:
+                new_packet = self.serial_buf[1:].find(self.PACKET_BOUNDARY) + 1
+                self.serial_buf = self.serial_buf[new_packet:]
+
+
+    def verbose_log(self, fmt_string, *fmt_args):
+        """Log a message if verbosity is enabled."""
+        if self.verbose:
+            print(fmt_string.format(*fmt_args))
+
+    async def process_messages(self):
+        """Process recieved messages."""
+        await self._ready.wait()
+        while not self.transport.is_closing():
+            msg = await self.read_queue.get()
+            msg_type = msg.get_message_id()
+            if msg_type not in self.response_map:
+                continue
+            self.response_map[msg_type](msg)
+
+    async def send_subscribed_params(self):
+        """Send values of subscribed parameters at a regular interval."""
+        await self._ready.wait()
+        device_id = hm.uid_to_device_id(self.uid)
+        while not self.transport.is_closing():
+            await asyncio.sleep(0.005, loop=self.event_loop)
+            if self.update_time != 0 and self.delay != 0:
+                if time.time() - self.update_time >= self.delay * 0.001:
+                    # If the time equal to the delay has elapsed since the previous device data,
+                    # send a device data with the device id
+                    # and the device's subscribed params and values
+                    data = []
+                    for param in self.subscribed_params:
+                        data.append((param, self.param_values[param]))
+                    hm.send(self.transport, hm.make_device_data(device_id, data))
+                    self.update_time = time.time()
+                    self.verbose_log("Regular data update sent from {}",
+                                     hm.uid_to_device_name(self.uid))
+
+    async def request_heartbeats(self):
+        """Request heartbeats on a regular basis."""
+        await self._ready.wait()
+        while not self.transport.is_closing():
+            await asyncio.sleep(self.HEARTBEAT_DELAY_MS/1000, loop=self.event_loop)
+            hm.send(self.transport, hm.make_heartbeat_request())
+
+    # pylint: disable=unused-argument
+    def _process_ping(self, msg):
+        """Respond to a ping packet."""
+        self.verbose_log("Ping received")
+        dev_id = hm.uid_to_device_id(self.uid)
+        hm.send(self.transport, hm.make_subscription_response(dev_id, [], 0, self.uid))
+
+    def _process_sub_request(self, msg):
+        """Respond to a subscription request with an appropriate response."""
+        self.update_time = time.time()
+        dev_id = hm.uid_to_device_id(self.uid)
+        self.verbose_log("Subscription request received")
+        params, delay = struct.unpack("<HH", msg.get_payload())
+        subscribed_params = hm.decode_params(dev_id, params)
+        hm.send(self.transport,
+                hm.make_subscription_response(dev_id, subscribed_params, delay, self.uid))
+        self.delay = delay
+        self.subscribed_params.update(set(subscribed_params))
+
+    def _process_device_read(self, msg):
+        self.verbose_log("Device read received")
+        device_id = hm.uid_to_device_id(self.uid)
+        # Send a device data with the requested param and value tuples
+        params, = struct.unpack("<H", msg.get_payload())
+        read_params = hm.decode_params(device_id, params)
+        read_data = []
+
+        for param in read_params:
+            if not (hm.readable(device_id, param) and param in self.param_values):
+                raise ValueError("Tried to read unreadable parameter {}".format(param))
+            read_data.append((param, self.param_values[param]))
+        hm.send(self.transport, hm.make_device_data(device_id, read_data))
+
+    def _process_device_write(self, msg):
+        # Write to requested parameters
+        # and return the values of the parameters written to using a device data
+        self.verbose_log("Device write received")
+        device_id = hm.uid_to_device_id(self.uid)
+        write_params_and_values = hm.decode_device_write(msg, device_id)
+
+        for (param, value) in write_params_and_values:
+            if not (hm.writable(device_id, param) and param in self.param_values):
+                raise ValueError("Tried to write read-only parameter: {}".format(param))
+            self.param_values[param] = value
+
+        updated_params = []
+        for (param, value) in write_params_and_values:
+            if hm.readable(device_id, param):
+                updated_params.append((param, value))
+        hm.send(self.transport, hm.make_device_data(device_id, updated_params))
+
+    def _process_disable(self, msg):
+        pass
+
+    def _process_heartbeat_response(self, msg):
+        pass
+
+    def connection_made(self, transport):
+        self.transport = transport
+        self._ready.set()
+
+    def connection_lost(self, exc):
+        """
+        If this happens, we're shutting down anyways,
+        so don't do anything.
+        """
+        pass
+
+
 def main():
     """
     Create virtual devices and send test data on them.
@@ -42,7 +207,6 @@ def main():
     device = args.device
     port = args.port
     verbose_log("Device {} on port {}", device, port)
-    conn = serial.Serial(port, 115200)
 
     for device_num in hm.DEVICES:
         if hm.DEVICES[device_num]["name"] == device:
@@ -51,119 +215,19 @@ def main():
     else:
         raise RuntimeError("Invalid Device Name!!!")
 
+
     year = 1
-    randomness = random.randint(0, 0xFFFFFFFFFFFFFFFF)
-    delay = 0
-    update_time = 0
+    randomness = random.getrandbits(64)
     uid = (device_id << 72) | (year << 64) | randomness
+    event_loop = asyncio.get_event_loop()
 
-    # Here, the parameters and values to be sent in device datas
-    # are set for each device type, the list of subscribed parameters is set to empty,
-    if device == "LimitSwitch":
-        subscribed_params = []
-        params_and_values = [("switch0", True), ("switch1", True),
-                             ("switch2", False), ("switch3", False)]
-    if device == "ServoControl":
-        subscribed_params = []
-        params_and_values = [("servo0", 2), ("enable0", True), ("servo1", 0),
-                             ("enable1", True), ("servo2", 5), ("enable2", True),
-                             ("servo3", 3), ("enable3", False)]
-    if device == "Potentiometer":
-        subscribed_params = []
-        params_and_values = [("pot0", 6.7), ("pot1", 5.5), ("pot2", 34.1), ("pot3", 0.15)]
-    if device == "YogiBear":
-        subscribed_params = []
-        params_and_values = [("enable", True), ("command_state", 1), ("duty_cycle", 1.0),
-                             ("pid_pos_setpoint", 2.0), ("pid_pos_kp", 3.0), ("pid_pos_ki", 4.0),
-                             ("pid_pos_kd", 5.0), ("pid_vel_setpoint", 6.0), ("pid_vel_kp", 7.0),
-                             ("pid_vel_ki", 8.0), ("pid_vel_kd", 9.0), ("current_thresh", 10.0),
-                             ("enc_pos", 11.0), ("enc_vel", 12.0), ("motor_current", 13.0),
-                             ("deadband", 0.5)]
-    if device == "RFID":
-        subscribed_params = []
-        params_and_values = [("id", 0), ("tag_detect", 0)]
-    if device == "BatteryBuzzer":
-        subscribed_params = []
-        params_and_values = [("is_unsafe", False), ("calibrated", True), ("v_cell1", 11.0),
-                             ("v_cell2", 11.0), ("v_cell3", 11.0), ("v_batt", 11.0),
-                             ("dv_cell2", 11.0), ("dv_cell3", 11.0)]
-    if device == "LineFollower":
-        subscribed_params = []
-        params_and_values = [("left", 1.0), ("center", 1.0), ("right", 1.0)]
+    def protocol_factory():
+        """Create a VirtualDevice with filled-in parameters."""
+        return VirtualDevice(uid, event_loop, args.verbose)
 
-    while True:
-        if update_time != 0 and delay != 0:
-            if time.time() - update_time >= delay * 0.001:
-                # If the time equal to the delay has elapsed since the previous device data,
-                # send a device data with the device id
-                # and the device's subscribed params and values
-                data = []
-                for data_tuple in params_and_values:
-                    if data_tuple[0] in subscribed_params and hm.readable(device_id, data_tuple[0]):
-                        data.append(data_tuple)
-                hm.send(conn, hm.make_device_data(device_id, data))
-                update_time = time.time()
-                verbose_log("Regular data update sent from {}", device)
-
-        msg = hm.read(conn)
-        if not msg:
-            time.sleep(.005)
-            continue
-        if msg.get_message_id() in [hm.MESSAGE_TYPES["SubscriptionRequest"]]:
-            # Update the delay, subscription time,
-            # and params, then send a subscription response
-            verbose_log("Subscription request received")
-            params, delay = struct.unpack("<HH", msg.get_payload())
-
-            subscribed_params = hm.decode_params(device_id, params)
-            hm.send(conn, hm.make_subscription_response(device_id, subscribed_params, delay, uid))
-            update_time = time.time()
-        if msg.get_message_id() in [hm.MESSAGE_TYPES["Ping"]]:
-            # Send a subscription response
-            verbose_log("Ping received")
-            hm.send(conn, hm.make_subscription_response(device_id, subscribed_params, delay, uid))
-        if msg.get_message_id() in [hm.MESSAGE_TYPES["DeviceRead"]]:
-            # Send a device data with the requested param and value tuples
-            verbose_log("Device read received")
-            params = struct.unpack("<H", msg.get_payload())
-            read_params = hm.decode_params(device_id, params)
-            read_data = []
-
-            for data_tuple in params_and_values:
-                if data_tuple[0] in read_params:
-                    if not hm.readable(device_id, data_tuple[0]):
-                        # Raise a syntax error if one of the values to be read is not readable
-                        raise SyntaxError("Attempted to read an unreadable value")
-                    read_data.append(data_tuple)
-            hm.send(conn, hm.make_device_data(device_id, read_data))
-
-        if msg.get_message_id() in [hm.MESSAGE_TYPES["DeviceWrite"]]:
-            # Write to requested parameters
-            # and return the values of the parameters written to using a device data
-            verbose_log("Device write received")
-            write_params_and_values = hm.decode_device_write(msg, device_id)
-            write_params = [param_val[0] for param_val in write_params_and_values]
-            value_types = [hm.PARAM_MAP[device_id][name][1] for name in write_params]
-
-            write_tuples = []
-            # pylint: disable=consider-using-enumerate
-            for index in range(len(write_params)):
-                write_tuples.append((write_params[index], write_params_and_values[index][1]))
-            for new_tuple in write_tuples:
-                if not hm.writable(device_id, new_tuple[0]):
-                    # Raise a syntax error if the value
-                    # that the message attempted to write to is not writable
-                    raise SyntaxError("Attempted to write to an unwritable value")
-                params_and_values[hm.PARAM_MAP[device_id][new_tuple[0]][0]] = new_tuple
-
-            # Send the written data, make sure you only send data for readable parameters
-            index = 0
-            while index < len(write_params):
-                if hm.readable(device_id, write_tuples[index][0]):
-                    index += 1
-                else:
-                    del write_tuples[index]
-            hm.send(conn, hm.make_device_data(device_id, write_tuples))
+    event_loop.create_task(serial_asyncio.create_serial_connection(event_loop, protocol_factory,
+                                                                   port, baudrate=115200))
+    event_loop.run_forever()
 
 
 if __name__ == "__main__":
